Stupid me, cast should be to int, not short (see last commit message, the message was right but the code was wrong)

+++ MapSimulation.java
@@ -70,8 +70,8 @@ class MapSimulation {
 	short getClosestWall(int[] coords, double heading) {
 		int[] c = coords;
 		while(mapItems.get(c[0]*c[1]).isEmpty()) { // while is empty (remove double negative, woot!)
-			c[0] += (short)Math.floor(Math.cos(direction));
-			c[1] += (short)Math.floor(Math.sin(direction));
+			c[0] += (int)Math.floor(Math.cos(direction));
+			c[1] += (int)Math.floor(Math.sin(direction));
 		}
 		// return the integer from the function sqrt(a²+b²) where a and b are the lengths of the vectors between coords and coords+c[0], and
 		// between coords and coords+c[1].  The result is the hypotenuse, according to Pythagorus' theorum.
@@ -82,8 +82,8 @@ class MapSimulation {
 	short getClosestSeekable(Item[] seeking, int[] coords, double direction) {
 		int[] c = coords;
 		while(mapItems.get(c[0]*c[1]).isEmpty()) { // while is empty (remove double negative, woot!)
-			c[0] += (short)Math.floor(Math.cos(direction));
-			c[1] += (short)Math.floor(Math.sin(direction));
+			c[0] += (int)Math.floor(Math.cos(direction));
+			c[1] += (int)Math.floor(Math.sin(direction));
 		}
 		// return the integer from the function sqrt(a²+b²) where a and b are the lengths of the vectors between coords and coords+c[0], and
 		// between coords and coords+c[1].  The result is the hypotenuse, according to Pythagorus' theorum.
